+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Instructions for adding the ads.txt file entry
 ### Fixed
 - Inform user while campaign has no active ads
 - Allow to add new advertisement to campaign without ads